requirements.txt
fastapi
uvicorn
jinja2
yfinance
sqlalchemy
tensorflow
keras
pillow
matplotlib
opencv-python
pandas
numpy
mtcnn
aiofiles
python-multipart
keras_facenet
pymongo